replaced int per size_t

+++ examples/main_sequential.cpp
@@ -92,7 +92,7 @@ int main(int argc, char **argv) {
 	//allocate couplings matrix
 	couplings = (size_t**)malloc(n_atomics * sizeof(size_t*));
 
-	for(int i = 0; i < n_atomics; i++){
+	for(size_t i = 0; i < n_atomics; i++){
 		couplings[i] = (size_t *)malloc(9 * sizeof(size_t));
 	}
 
@@ -100,7 +100,7 @@ int main(int argc, char **argv) {
 	int *n_couplings;
 
 	//allocate couplings matrix
-	n_couplings = (int *)malloc(n_atomics * sizeof(int));
+	n_couplings = (size_t *)malloc(n_atomics * sizeof(size_t));
 
 	//fill data structure for couplings
 	for(size_t i=0; i < n_atomics; i++){
